Remove suid/sgid checks from banner
Decided that it just doesn't make sense to try to cover this as part of
a threat model for this application. I can't point to another
application that explicitly tries to avoid being run suid/sguid. It
should suffice to leave that up to the environment. Documentation should
note that marking this application suid is a equivalent to allowing any
user with execute priveleges to become that user.